+++ bin/init.js
@@ -1,5 +1,6 @@
 #!/usr/bin/env node
 const util = require('yyl-util');
+const print = require('yyl-print');
 const iArgv = process.argv.splice(2);
 
 process.on('uncaughtException', (err) => {
@@ -9,4 +10,6 @@ process.on('uncaughtException', (err) => {
 const cmd = require('../index.js');
 const handle = iArgv[0];
 const iEnv = util.envPrase(iArgv);
-cmd.run(handle, iEnv).catch(() => {});
+cmd.run(handle, iEnv).catch((err) => {
+  print.log.error(err);
+});